requirements.txt
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.4
django-filter==2.4.0
django-oauth-toolkit==1.3.3
djangorestframework==3.12.2
et-xmlfile==1.0.1
idna==2.10
iniconfig==1.1.1
isort==5.6.4
jdcal==1.4.1
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
oauth2-provider==0.0
oauthlib==3.1.0
openpyxl==3.0.5
packaging==20.8
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.1
pytest-django==4.1.0
pytz==2020.5
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
